refactor(dev): extract transform utils from `codemod/`
