+++ packages/remix-dev/transform/plugins/recast.ts
@@ -2,13 +2,13 @@ import type Babel from "@babel/core";
 import type { File } from "@babel/types";
 import * as recast from "recast";
 
-import type { PluginObj } from "./babel";
+import type { Plugin } from "../plugin";
 
-function parse(
+let parse = (
   code: string,
   options: Babel.ParserOptions,
   parse: (code: string, options: Babel.ParserOptions) => File
-): File {
+): File => {
   return recast.parse(code, {
     parser: {
       parse(code: string) {
@@ -16,18 +16,18 @@ function parse(
       },
     },
   });
-}
+};
 
-function generate(ast: File): { code: string; map?: object } {
+let generate = (ast: File): { code: string; map?: object } => {
   return recast.print(ast);
-}
+};
 
 /**
  * Adapted from [@codemod/core](https://github.com/codemod-js/codemod/blob/5a9fc6968409613eefd87e646408c08b6dad0c40/packages/core/src/RecastPlugin.ts)
  */
-export default function (): PluginObj {
+export let plugin: Plugin = () => {
   return {
     parserOverride: parse,
     generatorOverride: generate,
   };
-}
+};
